ai/requirements.txt
numpy
pandas
scikit-learn
qdrant-client
boto3 
twelvelabs
pytube
flask
flask-cors
python-dotenv